+++ frontend/src/components/setting-device/create-button.tsx
@@ -20,7 +20,7 @@ import { useState, useEffect } from "react";
 import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import { SelectChangeEvent } from "@mui/material";
 import { ClimateDataResponse, CreateDeviceRequest } from "@/types/api";
-import { getClimateDatas } from "@/mocks/setting-device-api";
+import { getClimateDatas } from "@/features/api/mocks/setting-device-api";
 // import { getClimateDatas } from "@/features/api/climate-data/get-climate-data";
 import { createDevice } from "@/features/api/device/create-device";
