+++ frontend/src/components/workflow-control/sidebar.tsx
@@ -4,7 +4,7 @@ import { Box, Divider, Typography } from "@mui/material";
 import PlayCircleIcon from "@mui/icons-material/PlayCircle";
 import { useState, useEffect } from "react";
 import { DeviceResponse } from "@/types/api";
-import { getDevices } from "@/mocks/workflow-api";
+import { getDevices } from "@/features/api/mocks/workflow-api";
 
 export const Sidebar = () => {
   const [, setType] = useDnD();
