+++ frontend/src/features/api/mocks/setting-device-api.ts
